Move Jax single GPU example
