+++ docs/README.md
@@ -120,9 +120,9 @@ If you find any errors in the documentation, missing or unclear sections, or wou
         - [Software Frameworks](examples/frameworks/index.md)
             - [PyTorch Setup](examples/frameworks/pytorch_setup/index.md)
             - [Jax Setup](examples/frameworks/jax_setup/index.md)
-            - [Jax](examples/frameworks/jax/index.md)
         - [Distributed Training](examples/distributed/index.md)
             - [Single GPU Job](examples/distributed/single_gpu/index.md)
+                - [Single GPU Job with Jax](examples/distributed/single_gpu/single_gpu_jax/index.md)
             - [Multi-GPU Job](examples/distributed/multi_gpu/index.md)
             - [Multi-node Job](examples/distributed/multi_node/index.md)
         - [Good Practices](examples/good_practices/index.md)
