+++ docs/examples/distributed/single_gpu/single_gpu_jax/index.md
@@ -1,4 +1,4 @@
-# Jax
+# Single GPU Job with Jax
 
 ## Prerequisites
 
@@ -15,25 +15,25 @@ The full source code for this example is available on [the mila-docs GitHub repo
 **job.sh**
 
 ```bash
---8<-- "docs/examples/frameworks/jax/job.sh"
+--8<-- "docs/examples/distributed/single_gpu/single_gpu_jax/job.sh"
 ```
 
 **pyproject.toml**
 
 ```toml
---8<-- "docs/examples/frameworks/jax/pyproject.toml"
+--8<-- "docs/examples/distributed/single_gpu/single_gpu_jax/pyproject.toml"
 ```
 
 **main.py**
 
 ```python
---8<-- "docs/examples/frameworks/jax/main.py"
+--8<-- "docs/examples/distributed/single_gpu/single_gpu_jax/main.py"
 ```
 
 **model.py**
 
 ```python
---8<-- "docs/examples/frameworks/jax/model.py"
+--8<-- "docs/examples/distributed/single_gpu/single_gpu_jax/model.py"
 ```
 
 ## Running this example
