+++ docs/examples/distributed/single_gpu/single_gpu_jax/model.py
