+++ docs/examples/frameworks/index.md
@@ -9,6 +9,5 @@ These examples use [uv](../../technical_reference/general_theory/portability/#uv
 
 - [PyTorch Setup](pytorch_setup/index.md)
 - [Jax Setup](jax_setup/index.md)
-- [Jax](jax/index.md)
 
 * TensorFlow (Coming soon!)
